An example TM set as INITIALISATION.
Trying to use add_transition in ProB to define a program is more or
less impossible so hardcoding stuff in the abstract machine seems like
the only way to do this for now.

We also commented out add_transition to get a less bloated view of the
enabled operations in ProB. Similarly we set a lot of extra
constraints so there's only one possible way to SETUP_CONSTANS, and a
sane one too.